refactor(app): extract click and scroll-position helpers (phase 4c)
Move move_cursor_to_visible_area, calculate_max_scroll_position,
fold_toggle_line_at_screen_position, handle_editor_click,
handle_file_explorer_click out of input.rs into a new
app/click_handlers.rs sibling. ~513 lines.

Mouse-click dispatch (gutter/scrollbar/cursor placement/multi-cursor add)
and the small post-scroll cursor-fixup helpers. Pure click geometry
already lives in app/click_geometry.rs (phase 1g); these are the
side-effecting orchestrators that mutate buffer state, multi-cursor
arrays, and split view-states.

input.rs goes from 2304 -> 1791 lines (-513). 2026 tests still pass.
