+++ crates/fresh-editor/src/app/mod.rs
@@ -8,6 +8,7 @@ mod buffer_management;
 mod calibration_actions;
 pub mod calibration_wizard;
 mod click_geometry;
+mod click_handlers;
 mod clipboard;
 mod composite_buffer_actions;
 mod dabbrev_actions;
